user_service/requirements.txt
# User Service Requirements
# Core web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# HTTP client
httpx>=0.25.0

# File uploads
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0

# Environment variables (for migrations)
python-dotenv>=1.0.0

# Note: shared/ module is part of the monorepo and doesn't need separate installation

